flask_app/requirements.txt
Flask==2.0.2
flask-cors
requests==2.27.1
gunicorn
Pillow